place-recognition-system/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "place-recognition-system"
version = "0.1.0"
description = "Image-Based Place Recognition & Retrieval System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Arda Uçar"}
]
keywords = ["computer-vision", "place-recognition", "image-retrieval", "deep-learning"]

dependencies = [
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "rich>=13.0.0",
    "imagehash>=4.3.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
faiss = [
    "faiss-cpu>=1.7.4",  # For CPU-only systems
]
faiss-gpu = [
    "faiss-gpu>=1.7.4",  # For systems with CUDA support
]

[project.scripts]
place-recognition = "place_recognition.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
